Move exchange comparisons and etc to common/exchange
