+++ common/exchange/rates.go
@@ -1,29 +1,54 @@
-package txpool
+package exchange
 
 import (
+	"errors"
 	"fmt"
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/contracts"
 	"github.com/ethereum/go-ethereum/log"
 )
 
 var (
 	unitRate = big.NewRat(1, 1)
+	// ErrNonWhitelistedFeeCurrency is returned if the currency specified to use for the fees
+	// isn't one of the currencies whitelisted for that purpose.
+	ErrNonWhitelistedFeeCurrency = errors.New("non-whitelisted fee currency address")
 )
 
 // ConvertCurrency does an exchange conversion from currencyFrom to currencyTo of the value given.
 func ConvertCurrency(exchangeRates common.ExchangeRates, val1 *big.Int, currencyFrom *common.Address, currencyTo *common.Address) *big.Int {
-	goldAmount, err := contracts.ConvertCurrencyToGold(exchangeRates, val1, currencyFrom)
+	goldAmount, err := ConvertCurrencyToGold(exchangeRates, val1, currencyFrom)
 	if err != nil {
 		log.Error("Error trying to convert from currency to gold.", "value", val1, "fromCurrency", currencyFrom.Hex())
 	}
-	toAmount, err := contracts.ConvertGoldToCurrency(exchangeRates, currencyTo, goldAmount)
+	toAmount, err := ConvertGoldToCurrency(exchangeRates, currencyTo, goldAmount)
 	if err != nil {
 		log.Error("Error trying to convert from gold to currency.", "value", goldAmount, "toCurrency", currencyTo.Hex())
 	}
 	return toAmount
+}
+
+func ConvertCurrencyToGold(exchangeRates common.ExchangeRates, currencyAmount *big.Int, feeCurrency *common.Address) (*big.Int, error) {
+	if feeCurrency == nil {
+		return currencyAmount, nil
+	}
+	exchangeRate, ok := exchangeRates[*feeCurrency]
+	if !ok {
+		return nil, ErrNonWhitelistedFeeCurrency
+	}
+	return new(big.Int).Div(new(big.Int).Mul(currencyAmount, exchangeRate.Denom()), exchangeRate.Num()), nil
+}
+
+func ConvertGoldToCurrency(exchangeRates common.ExchangeRates, feeCurrency *common.Address, goldAmount *big.Int) (*big.Int, error) {
+	if feeCurrency == nil {
+		return goldAmount, nil
+	}
+	exchangeRate, ok := exchangeRates[*feeCurrency]
+	if !ok {
+		return nil, ErrNonWhitelistedFeeCurrency
+	}
+	return new(big.Int).Div(new(big.Int).Mul(goldAmount, exchangeRate.Num()), exchangeRate.Denom()), nil
 }
 
 func getRate(exchangeRates common.ExchangeRates, feeCurrency *common.Address) (*big.Rat, error) {
