+++ common/exchange/rates_test.go
@@ -1,4 +1,4 @@
-package txpool
+package exchange
 
 import (
 	"math/big"
